added loot component (takes an id of looter) added inventory copy function to Inventory.scala moved attack code from collision system to attackSystem created attack system and added to world factory did more work on statuseffects (still not working)
