+++ src/main/scala/ayai/components/Inventory.scala
@@ -47,4 +47,12 @@ case class Inventory (inventory: ArrayBuffer[Item] = new ArrayBuffer[Item]()) ex
     return totalWeight
   }
 
+  def copy(): Inventory = {
+    var copyInventory: ArrayBuffer[Item] = new ArrayBuffer[Item]()
+    for(item <- inventory) {
+      copyInventory += item.copy
+    }
+    new Inventory(copyInventory)
+  }
+
 }
