+++ src/main/scala/ayai/factories/WorldFactory.scala
@@ -19,6 +19,7 @@ class WorldFactory(networkSystem: ActorSystem) {
     world.addSystem(RoomChangingSystem(networkSystem))
     world.addSystem(NetworkingSystem(networkSystem))
     world.addSystem(CollisionSystem(networkSystem))
+    world.addSystem(AttackSystem(networkSystem))
     val serializer = networkSystem.actorOf(Props(GameStateSerializer(world)), s"Serializer$name")
     val nmProcessor = networkSystem.actorOf(Props(MessageProcessorSupervisor(world)), name=s"MProcessor$name")
